Add customer bean
